Add AST lint + live-Python rot guard for the FMA class; fix one more site
The AST-based lint (internal/jobs/fma_lint_test.go, go/packages + go/types
over internal/jobs/...) found one unguarded site the manual sweep missed:
computePipelineStability's weights[i] = 1.0 + float64(i)*0.5
(testops_risk_native_clickhouse.go), matching CPython's per-element list
comprehension. Fixed with the same explicit-conversion pattern; verified
analytically that this specific site cannot actually diverge (multiplying
by 0.5 is exact for any representable float, so there is no rounding to
lose either way) but applies the guard for consistency with every other
site and so the lint stays green.

The lint type-checks every non-test .go file under internal/jobs/... and
flags only genuine float32/float64 x*y +/- z (never a superficially
similar int or time.Duration expression). It is a CI-arch-independence
delivery in place of an arm64 GitHub Actions matrix entry: this lane's PR
cannot merge a .github/workflows/* change (gh token lacks the workflow
scope), and the lint needs no runner architecture at all -- it flags the
pattern regardless of which CPU it runs on. It cannot catch the second,
subtler exposure this ticket found (fusion across statements via
rematerialization around a function call) -- that class is covered by the
bit-pattern golden tests instead; see the lint's own doc comment.

Also registers the CHAOS-4818 FMA golden's live-Python rot guard in
ci/check_go.sh live-python-oracles, in the same shape as the other three
numerical package goldens (own proof marker, verified independently).

+++ internal/jobs/metrics/daily/testops_risk_native_clickhouse.go
@@ -818,7 +818,11 @@ func computePipelineStability(repoID uuid.UUID, day time.Time, dayEntries []test
 	weights := make([]float64, n)
 	totalWeight := 0.0
 	for i := range dayEntries {
-		weights[i] = 1.0 + float64(i)*0.5
+		// float64(...) is load-bearing (CHAOS-4818): Go may otherwise fuse
+		// this into one FMA on arm64, rounding once where CPython's
+		// `weights = [1.0 + i * 0.5 for i in range(n)]` rounds the multiply
+		// and the add separately.
+		weights[i] = 1.0 + float64(float64(i)*0.5)
 		totalWeight += weights[i]
 	}
 	successRate7d := 0.0
